Added StorageManager class to handle all read/write of data and make Storage more OOP
