+++ .gitignore
@@ -15,3 +15,4 @@ bin/
 
 /duke.txt
 /wallet.txt
+/data
