+++ src/main/java/wallet/logic/LogicManager.java
@@ -10,22 +10,24 @@ import wallet.model.record.RecordList;
 import wallet.storage.Storage;
 import wallet.model.task.ScheduleList;
 import wallet.model.task.TaskList;
+import wallet.storage.StorageManager;
+import wallet.storage.TaskStorage;
 
 /**
  * The LogicManager Class handles the logic of Wallet.
  */
 public class LogicManager {
-    private final Storage storage;
+    private final StorageManager storageManager;
     private final ParserManager parserManager;
     private final Wallet wallet;
 
     /**
      * Constructs a LogicManager object.
      */
-    public LogicManager(Storage storage) {
-        this.storage = storage;
+    public LogicManager() {
+        this.storageManager = new StorageManager();
         this.wallet = new Wallet(new RecordList(), new ExpenseList(), new ContactList(),
-                new TaskList(storage.loadFile()), new ScheduleList(), new LoanList());
+                new TaskList(storageManager.loadTask()), new ScheduleList(), new LoanList());
         this.parserManager = new ParserManager();
     }
 
@@ -38,7 +40,7 @@ public class LogicManager {
         boolean isExit = false;
         try {
             Command command = parserManager.parseCommand(fullCommand);
-            isExit = command.execute(wallet, storage);
+            isExit = command.execute(wallet, storageManager);
         } catch (Exception e) {
             e.printStackTrace();
             System.out.println("Error encountered while executing command.");
