+++ src/main/java/wallet/logic/command/Command.java
@@ -1,7 +1,7 @@
 package wallet.logic.command;
 
 import wallet.model.Wallet;
-import wallet.storage.Storage;
+import wallet.storage.StorageManager;
 
 public abstract class Command {
 
@@ -9,8 +9,8 @@ public abstract class Command {
      * Returns a boolean indicating whether the user wants to continue running the application.
      *
      * @param wallet The Wallet object.
-     * @param storage The Storage object.
+     * @param storageManager The StorageManager object.
      * @return A boolean indicating whether the user wants to continue running the application.
      */
-    public abstract boolean execute(Wallet wallet, Storage storage);
+    public abstract boolean execute(Wallet wallet, StorageManager storageManager);
 }
