+++ src/main/java/wallet/logic/command/DeleteCommand.java
@@ -1,8 +1,8 @@
 package wallet.logic.command;
 
 import wallet.model.Wallet;
-import wallet.storage.Storage;
 import wallet.model.task.Task;
+import wallet.storage.StorageManager;
 
 /**
  * DeleteCommand class handles any command that involves deletion of Record objects.
@@ -29,16 +29,16 @@ public class DeleteCommand extends Command {
      * Deletes the specific Record object and returns false.
      *
      * @param wallet The Wallet object.
-     * @param storage The Storage object.
+     * @param storageManager The StorageManager object.
      * @return A boolean which indicates whether program terminates.
      */
     @Override
-    public boolean execute(Wallet wallet, Storage storage) {
+    public boolean execute(Wallet wallet, StorageManager storageManager) {
         switch (object) {
         case "task":
             Task task = wallet.getTaskList().getTask(index);
             wallet.getTaskList().deleteTask(index);
-            storage.removeTask(wallet.getTaskList().getTaskList(), index);
+            storageManager.deleteTask(wallet.getTaskList().getTaskList(), index);
             System.out.println(MESSAGE_SUCCESS_DELETE_TASK);
             System.out.println(task.toString());
             break;
