+++ src/main/java/wallet/logic/command/EditCommand.java
@@ -1,7 +1,7 @@
 package wallet.logic.command;
 
 import wallet.model.Wallet;
-import wallet.storage.Storage;
+import wallet.storage.StorageManager;
 
 /**
  * EditCommand Class deals with commands that involves
@@ -12,7 +12,7 @@ public class EditCommand extends Command {
     public static final String COMMAND_WORD = "edit";
 
     @Override
-    public boolean execute(Wallet wallet, Storage storage) {
+    public boolean execute(Wallet wallet, StorageManager storageManager) {
         return false;
     }
 }
