+++ src/main/java/wallet/logic/command/ExitCommand.java
@@ -1,7 +1,7 @@
 package wallet.logic.command;
 
 import wallet.model.Wallet;
-import wallet.storage.Storage;
+import wallet.storage.StorageManager;
 
 /**
  * The ExitCommand Class deals with the 'bye' command.
@@ -13,11 +13,11 @@ public class ExitCommand extends Command {
      * Returns true to indicate to Main to terminate the application.
      *
      * @param wallet The Wallet object.
-     * @param storage The Storage object.
+     * @param storageManager The StorageManager object.
      * @return True. (Always)
      */
     @Override
-    public boolean execute(Wallet wallet, Storage storage) {
+    public boolean execute(Wallet wallet, StorageManager storageManager) {
         return true;
     }
 }
