+++ src/main/java/wallet/logic/command/HelpCommand.java
@@ -1,7 +1,7 @@
 package wallet.logic.command;
 
 import wallet.model.Wallet;
-import wallet.storage.Storage;
+import wallet.storage.StorageManager;
 
 /**
  * The HelpCommand Class shows users what the valid commands are.
@@ -13,11 +13,11 @@ public class HelpCommand extends Command {
      * Shows a list of valid commands to the user and returns false.
      *
      * @param wallet The Wallet object.
-     * @param storage The Storage object.
+     * @param storageManager The StorageManager object.
      * @return false.
      */
     @Override
-    public boolean execute(Wallet wallet, Storage storage) {
+    public boolean execute(Wallet wallet, StorageManager storageManager) {
         return false;
     }
 }
