+++ src/main/java/wallet/logic/command/ListCommand.java
@@ -5,7 +5,7 @@ import wallet.model.contact.Contact;
 import wallet.model.record.Expense;
 import wallet.model.record.Loan;
 import wallet.model.task.Task;
-import wallet.storage.Storage;
+import wallet.storage.StorageManager;
 
 /**
  * The ListCommand Class handles all list commands.
@@ -40,11 +40,11 @@ public class ListCommand extends Command {
      * Lists the Record objects in any list and returns false.
      *
      * @param wallet The Wallet object.
-     * @param storage The Storage object.
+     * @param storageManager The StorageManager object.
      * @return False.
      */
     @Override
-    public boolean execute(Wallet wallet, Storage storage) {
+    public boolean execute(Wallet wallet, StorageManager storageManager) {
         boolean isListAll = false;
         int counter;
         switch (record) {
